backend-python/requirements.txt
requests
langchain
langchain_community
langchain_openai
python-dotenv